Remove experiment folder

+++ src/main/java/MultiIheritance/FlatFigure.java
@@ -1,5 +1,5 @@
 package MultiIheritance;
 
 public interface FlatFigure {
-    public int area();
+    public double area();
 }
