+++ src/main/java/MultiIheritance/Rectangle.java
@@ -15,14 +15,13 @@ public class Rectangle implements Drawable, FlatFigure {
         for (int i = 0; i < height; i++){
             for(int j = 0; j < width; j++){
                 System.out.print("*");
-
             }
             System.out.println();
         }
     }
 
     @Override
-    public int area() {
+    public double area() {
         return height * width;
     }
 }
